firestore 쓰임에 따라 파일명 변경
firestore -> recruit
